.claude/worktrees/fervent-khayyam-60b5a6/src/backend/azure_functions/strategic_autonomy_timer/requirements.txt
azure-functions
azure-identity
azure-servicebus
python-dotenv
httpx
